first commit towards graphs
