+++ advanced/sr_gui_fingertip_visualization/sr_gui_biotac_plugin.xml
@@ -1,7 +1,7 @@
 <library path="src">
   <class name="SrFingertipVisualizer" type="sr_gui_fingertip_visualization.sr_gui_fingertip_visualization.SrFingertipVisualizer" base_class_type="rqt_gui_py::Plugin">
     <description>
-      A plugin for visualising data from Syntouch Biotac finger tip sensors
+      A plugin for visualising data from tactile fingertips
     </description>
     <qtgui>
       <group>
@@ -9,7 +9,7 @@
       </group>
       <label>Fingertip visualization</label>
       <icon type="file">resource/hand.png</icon>
-      <statustip>A plugin for visualising data from Syntouch Biotac finger tip sensors</statustip>
+      <statustip>A plugin for visualising data from tactile fingertips</statustip>
     </qtgui>
   </class>
 </library>
