+++ advanced/sr_gui_fingertip_visualization/src/sr_gui_fingertip_visualization/sr_gui_fingertip_visualization.py
@@ -33,26 +33,31 @@ from python_qt_binding.QtWidgets import (
     QHBoxLayout,
     QPushButton,
     QComboBox,
-    QMessageBox
+    QMessageBox,
+    QLabel
 )
 
-from sr_gui_fingertip_visualization.tab_layouts import (
-    PSTVisualizationTab,
-    BiotacVisualizationTab,
-    GenericTabLayout,
-    VisualizationTab
-)
-
+from sr_gui_fingertip_visualization.visual_tab_layouts import VisualizationTab
+from sr_gui_fingertip_visualization.graph_tab_layouts import GraphTab
 
 class SrFingertipVisualizer(Plugin):
     TITLE = "Fingertip Visualizer"
 
     def __init__(self, context):
         super().__init__(context)
+        self._detect_hand_and_tactile_type()
         self.context = context
-        self.init_ui()
+        self._init_ui()
 
-    def init_ui(self):
+    def _detect_hand_and_tactile_type(self):
+        type_right = rostopic.get_topic_type("/rh/tactile")
+        type_left = rostopic.get_topic_type("/lh/tactile")
+
+        self._hand_ids = [i[1].split('/')[1] for i in [type_right, type_left] if i[1]]
+        self._types = [i[0].split('/')[1] for i in [type_right, type_left] if i[1]]
+        self._tactile_topics = dict(zip(self._hand_ids, self._types))
+
+    def _init_ui(self):
         self._widget = QWidget()
         self.main_layout = QVBoxLayout()
         self.main_layout.setContentsMargins(0, 0, 0, 0)
@@ -70,25 +75,28 @@ class SrFingertipVisualizer(Plugin):
         self.information_btn = QPushButton("Info")
         self.main_layout.addWidget(self.information_btn, alignment=Qt.AlignRight)
 
-        # Initialize tabs
         self.tab_container = QTabWidget()
-        self.main_layout.addWidget(self.tab_container)
-
-        # Create tabs
-        self.create_tab("Visualizer")
-        self.create_tab("Graphs")
-
         self.tab_container.currentChanged.connect(self.tab_changed)
         self.information_btn.clicked.connect(self.display_information)
+
+        if not list(self._tactile_topics.keys()):
+            label = QLabel("No tactiles", self.tab_container)
+            label.setSizePolicy(1,1)            
+            self.main_layout.addWidget(label, alignment=Qt.AlignCenter)
+        else:
+            self.create_tab("Visualizer")
+            self.create_tab("Graphs")
+            self.main_layout.addWidget(self.tab_container)
 
     def create_tab(self, tab_name):
 
         if tab_name == "Visualizer":
-            self.tab_created = VisualizationTab(self._widget)
+            self.tab_created = VisualizationTab(self._widget, self._tactile_topics)
             self.tab_container.addTab(self.tab_created, tab_name)
 
         elif tab_name == "Graphs":
-            pass  # For now
+            self.tab_created = GraphTab(self._widget, self._tactile_topics)
+            self.tab_container.addTab(self.tab_created, tab_name)
 
     def tab_changed(self, index):
         pass
